refactor: replace ThemeToggle with ThemeSwitch in Navbar, Sidebar, and update main styles for dark mode
